proof: refine instantaneous SEA3 process cells
